etl_pipeline/requirements.txt
python-dotenv
psycopg2
requests
pandas